Move addListener up to testpageloader
